Adding ProxyService internal implementation. Renaming permission 'log_monitor' -> 'view_logs'. Adding 'view_logs' permission to site_admin role
