+++ src/main/java/org/craftercms/studio/permissions/StudioPermissionsConstants.java
@@ -79,7 +79,7 @@ public final class StudioPermissionsConstants {
 
     public static final String PERMISSION_CONTENT_SEARCH = "content_search";
 
-    public static final String PERMISSION_LOG_MONITOR = "log_monitor";
+    public static final String PERMISSION_VIEW_LOGS = "view_logs";
 
     private StudioPermissionsConstants() {
     }
