+++ src/main/resources/crafter/studio/upgrade/4.0.x/4.0.1.5/system/global-permission-mappings-config.xslt
@@ -39,9 +39,9 @@
         <xsl:copy>
             <xsl:copy-of select="@*"/>
             <xsl:apply-templates/>
-            <xsl:if test="not(permission = 'log_monitor')">
+            <xsl:if test="not(permission = 'view_logs')">
                 <xsl:element name="permission">
-                    <xsl:text>log_monitor</xsl:text>
+                    <xsl:text>view_logs</xsl:text>
                 </xsl:element>
                 <xsl:text>&#10;</xsl:text>
             </xsl:if>
