+++ src/main/webapp/repo-bootstrap/global/blueprints/2000_headless_store/config/studio/permission-mappings-config.xml
@@ -230,7 +230,7 @@
                 <permission>encryption_tool</permission>
                 <permission>get_children</permission>
                 <permission>publish_status</permission>
-                <permission>log_monitor</permission>
+                <permission>view_logs</permission>
             </allowed-permissions>
         </rule>
         <rule regex="^/(?!site/website/index\.xml)(.*)">
@@ -311,6 +311,7 @@
                 <permission>unlock_repository</permission>
                 <permission>item_unlock</permission>
                 <permission>repair_repository</permission>
+                <permission>view_logs</permission>
             </allowed-permissions>
         </rule>
         <rule regex="^/(?!site/website/index\.xml)(.*)">
