+++ src/main/webapp/repo-bootstrap/global/blueprints/4000_empty/config/studio/permission-mappings-config.xml
@@ -230,7 +230,7 @@
                 <permission>encryption_tool</permission>
                 <permission>get_children</permission>
                 <permission>publish_status</permission>
-                <permission>log_monitor</permission>
+                <permission>view_logs</permission>
             </allowed-permissions>
         </rule>
         <rule regex="^/(?!site/website/index\.xml)(.*)">
@@ -312,6 +312,7 @@
                 <permission>unlock_repository</permission>
                 <permission>item_unlock</permission>
                 <permission>repair_repository</permission>
+                <permission>view_logs</permission>
             </allowed-permissions>
         </rule>
         <rule regex="^/(?!site/website/index\.xml)(.*)">
