+++ src/test/resources/crafter/studio/upgrade/xslt/permission-mappings-config/permission-mappings-config-v4.0.7/expected.xml
@@ -195,7 +195,7 @@
                 <permission>remove_plugins</permission>
                 <permission>get_children</permission>
                 <permission>content_search</permission>
-                <permission>log_monitor</permission>
+                <permission>view_logs</permission>
             </allowed-permissions>
         </rule>
     </role>
@@ -283,6 +283,7 @@
                 <permission>publish_status</permission>
                 <permission>item_unlock</permission>
                 <permission>content_search</permission>
+                <permission>view_logs</permission>
             </allowed-permissions>
         </rule>
     </role>
